refactor: ajustes com o novo tabber
