--- src/components/Tabber/PostModes/PostModes.module.scss
@@ -1,76 +1,0 @@
-@use '~styles/global.scss';
-
-.postModesHeader {
-  max-width: 44rem;
-  overflow-x: auto;
-
-  margin-bottom: 1.6rem;
-
-  scroll-behavior: smooth;
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-
-.postModesWrapper {
-  width: fit-content;
-
-  display: flex;
-}
-
-.postModeTitle {
-  color: global.$primaryGray;
-  font-size: 1.4rem;
-  font-weight: 500;
-
-  padding: 1.2rem;
-
-  cursor: pointer;
-
-  &.active {
-    color: global.$secondaryPurple;
-  }
-}
-
-.selectPostMode {
-  display: flex;
-  gap: 0.4rem;
-
-  align-items: center;
-
-  color: global.$tertiaryGray;
-
-  &.active {
-    [type='checkbox'] {
-      border-color: global.$secondaryPurple;
-    }
-  }
-}
-
-@include from905 {
-  .postModesHeader {
-    border: 0.1rem solid global.$primaryGray;
-
-    background-color: global.$quaternaryWhite;
-    border-radius: 3rem;
-  }
-
-  .postModeTitle {
-    padding: 1.2rem;
-  }
-
-  .selectPostMode {
-    border: 0.2rem solid transparent;
-
-    &.active {
-      border-color: global.$secondaryPurple;
-
-      background-color: global.$primaryWhite;
-      border-radius: 10rem;
-    }
-  }
-}
-
