+++ src/pages/home/components/Tabber/PostModes/PostModes.tsx
@@ -2,55 +2,76 @@
 
 import classNames from 'classnames';
 
-import {
-  PostMode,
-  SocialMedia,
-} from '~services/api/social-media/social-media.types';
+import { PostMode as PostModeTypes } from '~services/api/social-media/social-media.types';
 import { useSocialMediaStore } from '~stores/useSocialMediaStore/useSocialMediaStore';
-
-import { Checkbox } from '~components/Checkbox/Checkbox';
 
 import scss from './PostModes.module.scss';
 
-import { PostModesProps } from './PostModes.types';
+import { PostMode } from './PostModes.components';
+import { PostModesProps, SelectedPostMode } from './PostModes.types';
 
 function PostModes(props: PostModesProps): ReactNode {
   const { socialMedias } = useSocialMediaStore();
+  const socialMedia = socialMedias.get(props.socialMediaId);
+  const currentTab = props.postId;
 
-  const postModes = socialMedias.get(props.socialMediaId)?.postModes;
-
-  const [selectedPostModes, setSelectedPostModes] = useState<Set<string>>(
-    new Set()
+  const [selectedPostModes, setSelectedPostModes] = useState<SelectedPostMode>(
+    {}
   );
 
-  const postModeClasses = (
-    postModeId: SocialMedia['postModes'][number]['id']
-  ): string =>
+  const addPostMode = (postModeId: PostModeTypes['id']): SelectedPostMode => {
+    const previousModes = selectedPostModes[currentTab] ?? [];
+    const updatedModes = [...previousModes, postModeId];
+    return {
+      ...selectedPostModes,
+      [currentTab]: updatedModes,
+    };
+  };
+
+  const removePostMode = (
+    postModeId: PostModeTypes['id']
+  ): SelectedPostMode => {
+    const previousModes = selectedPostModes[currentTab] ?? [];
+    const updatedModes = previousModes.filter((id) => id !== postModeId);
+    return {
+      ...selectedPostModes,
+      [currentTab]: updatedModes,
+    };
+  };
+
+  const onChangeCheckBox = (
+    postModeId: PostModeTypes['id'],
+    isChecked: boolean
+  ): void => {
+    if (isChecked) setSelectedPostModes(addPostMode(postModeId));
+    else setSelectedPostModes(removePostMode(postModeId));
+  };
+
+  const isChecked = (postModeId: PostModeTypes['id']): boolean =>
+    Boolean(selectedPostModes[currentTab]?.includes(postModeId));
+
+  const postModeClasses = (postModeId: PostModeTypes['id']): string =>
     classNames({
       [scss.active]: props.postModeId === postModeId,
       [scss.postModeTitle]: true,
+      [scss.selectPostMode]: true,
     });
 
-  const changePostMode = (postMode: PostMode): void => {
-    props.changePostModeId(postMode.id);
-  };
-
-  const renderPostMode = (postMode: PostMode): ReactNode => (
-    <button
-      className={`${scss.selectPostMode} ${postModeClasses(postMode.id)}`}
-      key={postMode.id}
-      onClick={() => changePostMode(postMode)}
-    >
-      <Checkbox
-        checked={selectedPostModes.has(postMode.id)}
-        onChange={(checked) => onChangeCheckBox(postMode.id, checked)}
-      />
-      {postMode.name}
-    </button>
-  );
-
   return (
-    <div className={scss.postModesHeader}>{postModes?.map(renderPostMode)}</div>
+    <div className={scss.postModesHeader}>
+      <div className={scss.postModesWrapper}>
+        {socialMedia?.postModes.map((postMode) => (
+          <PostMode
+            changeCheckBox={onChangeCheckBox}
+            changePostMode={props.changePostModeId}
+            isChecked={isChecked}
+            key={postMode.id}
+            postMode={postMode}
+            postModeClasses={postModeClasses}
+          />
+        ))}
+      </div>
+    </div>
   );
 }
